fix: moved handledrop to burger-constructor. Bug #4
